move extractor em dto
